Forget to change something to int in last commit, fixed forgot to break out when bounty amount was too low. I really need to test stuff before I push it

+++ src/main/java/net/conriot/prison/command/bounty/BountyListCommand.java
@@ -86,7 +86,7 @@ public class BountyListCommand extends AbstractCommand
 	
 	private static class NameComparableBounty implements Comparable<NameComparableBounty> 
 	{
-        public final double bounty;
+        public final int bounty;
         public final String name;
         private final String lowerName;
